+++ apps/dashboard/components/DatasourceForms/TextForm.tsx
@@ -3,39 +3,20 @@ import FormLabel from '@mui/joy/FormLabel';
 import Textarea from '@mui/joy/Textarea';
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
-import { z } from 'zod';
 
-import Input from '@app/components/Input';
-
-import { UpsertDatasourceSchema } from '@chaindesk/lib/types/models';
+import { DatasourceSchema, DatasourceText } from '@chaindesk/lib/types/models';
 import { DatasourceType } from '@chaindesk/prisma';
 
 import Base from './Base';
 import type { DatasourceFormProps } from './types';
 
-type Props = DatasourceFormProps & {};
-
-export const TextSourceSchema = UpsertDatasourceSchema.extend({
-  config: z
-    .object({
-      source_url: z.string().trim().optional(),
-    })
-    .optional(),
-});
+type Props = DatasourceFormProps<DatasourceText> & {};
 
 function Nested() {
-  const { control, register } =
-    useFormContext<z.infer<typeof TextSourceSchema>>();
+  const { control, register } = useFormContext<DatasourceText>();
 
   return (
     <>
-      <Input
-        label="Source URL (optional)"
-        control={control as any}
-        placeholder="https://news.ycombinator.com"
-        {...register('config.source_url')}
-      />
-
       <FormControl>
         <FormLabel>Text</FormLabel>
         <Textarea maxRows={21} minRows={4} {...register('datasourceText')} />
@@ -49,7 +30,7 @@ export default function TextForm(props: Props) {
 
   return (
     <Base
-      schema={TextSourceSchema}
+      schema={DatasourceSchema}
       {...rest}
       defaultValues={{
         ...props.defaultValues!,