+++ apps/dashboard/pages/api/agents/[id]/query.tsx
@@ -76,7 +76,7 @@ export const chatAgentRequest = async (
             },
             include: {
               messages: {
-                take: -4,
+                take: -24,
                 orderBy: {
                   createdAt: 'asc',
                 },