+++ packages/lib/loaders/text.ts
@@ -1,8 +1,10 @@
 import { AppDocument } from '@chaindesk/lib/types/document';
 
+import { DatasourceText } from '../types/models';
+
 import { DatasourceLoaderBase } from './base';
 
-export class TextLoader extends DatasourceLoaderBase {
+export class TextLoader extends DatasourceLoaderBase<DatasourceText> {
   async getSize(text: string) {
     return new Blob([text]).size;
   }
@@ -16,9 +18,9 @@ export class TextLoader extends DatasourceLoaderBase {
           datasource_id: this.datasource.id,
           datasource_name: this.datasource.name,
           datasource_type: this.datasource.type,
-          source_url: (this.datasource?.config as any)?.source_url,
-          custom_id: (this.datasource?.config as any)?.custom_id,
-          tags: [],
+          source_url: this.datasource?.config?.source_url!,
+          custom_id: this.datasource?.config?.custom_id,
+          tags: this.datasource?.config?.tags || [],
         },
       }),
     ];